rename to match new name of kata
