+++ csharp/BuildPipeline.sln
@@ -1,8 +1,8 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UntangledConditionals", "UntangledConditionals\UntangledConditionals.csproj", "{0AF5E38A-1C4B-4357-8D63-430A4862BF32}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BuildPipeline", "BuildPipeline\BuildPipeline.csproj", "{0AF5E38A-1C4B-4357-8D63-430A4862BF32}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UntangledConditionalsTest", "UntangledConditionalsTest\UntangledConditionalsTest.csproj", "{37839DFE-A59F-4A37-9BAE-81D5B9DC3AA5}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BuildPipelineTests", "BuildPipelineTests\BuildPipelineTests.csproj", "{37839DFE-A59F-4A37-9BAE-81D5B9DC3AA5}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
