+++ csharp/BuildPipeline/BuildPipeline.csproj
@@ -2,6 +2,7 @@
 
     <PropertyGroup>
         <TargetFramework>netcoreapp3.1</TargetFramework>
+        <RootNamespace>UntangledConditionals</RootNamespace>
     </PropertyGroup>
 
 </Project>
